+++ src/FolkerKinzel.MimeTypes/md/ReleaseNotes.md
@@ -1,4 +1,7 @@
-﻿- Dependency update
+﻿- **Important update:**
+    Fixes a bug in `MimeTypeInfo`: A plus sign in a not URL-encoded continuation of a URL-encoded splitted parameter could have accidentally been changed to a SPACE sign by the decoder.
+
+- Dependency update
 
 
 >**Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and check the "Allow" checkbox - if it is present - in the lower right corner of the General tab in the Properties dialog.